+++ doc/legal/legal.md
@@ -19,7 +19,7 @@ See `epl-1.0.txt`, `gpl-2.txt`, `lgpl-2.1.txt`.
 The standard implementation of Ruby is MRI. TruffleRuby contains code from MRI
 version 2.3.6, including: 
 - the standard library in `lib/mri`, 
-- Ruby C extension API in `lib/cext` and `src/main/c/cext`, 
+- Ruby C extension API in `lib/cext/include` and `src/main/c/cext`, 
 - OpenSSL C extension in `src/main/c/openssl`,
 - and zlib C extension in `src/main/c/zlib`.
