+++ src/main/c/Makefile
@@ -13,7 +13,7 @@ TRUFFLE_POSIX := $(ROOT)/lib/cext/truffleposix.$(DLEXT)
 TRUFFLE_MOCK := $(ROOT)/lib/cext/sulongmock.bc
 RUBY_LIBRARY := $(ROOT)/lib/cext/ruby.su
 
-RUBY_HEADERS := $(wildcard $(ROOT)/lib/cext/*.h) $(wildcard $(ROOT)/lib/cext/ruby/*.h)
+RUBY_HEADERS := $(wildcard $(ROOT)/lib/cext/include/*.h) $(wildcard $(ROOT)/lib/cext/include/ruby/*.h)
 RBCONFIG := $(ROOT)/lib/truffle/rbconfig.rb
 RBCONFIG_MKMF := $(ROOT)/lib/truffle/rbconfig-for-mkmf.rb
 PREPROCESS := $(ROOT)/lib/cext/preprocess.rb $(wildcard $(ROOT)/lib/cext/patches/*.rb)
